+++ x/oracle/keeper/task_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"crypto/sha256"
 	"testing"
 	"time"
 
@@ -308,4 +309,63 @@ func TestTaskAboveThreshold(t *testing.T) {
 	require.Nil(t, err)
 	require.Equal(t, addrs[2].String(), operator2.Address)
 	require.Equal(t, sdk.Coins{sdk.NewInt64Coin("uctk", 41666)}, operator2.AccumulatedRewards)
+}
+
+func TestTxTaskBasic(t *testing.T) {
+	app := shentuapp.Setup(false)
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{Time: time.Now().UTC()})
+	addrs := shentuapp.AddTestAddrs(app, ctx, 1, sdk.NewInt(80000*1e6))
+	ok := app.OracleKeeper
+
+	bounty := sdk.Coins{sdk.NewInt64Coin("uctk", 100000)}
+	businessHash := sha256.Sum256([]byte("hello"))
+
+	expiration1 := time.Now().Add(time.Hour).UTC()
+
+	txTask := types.TxTask{
+		Creator:   addrs[0].String(),
+		TxHash:    businessHash[:],
+		Bounty:    bounty,
+		ValidTime: expiration1,
+		Status:    types.TaskStatusPending,
+	}
+	require.NoError(t, ok.CreateTask(ctx, addrs[0], &txTask))
+
+	task1, err := ok.GetTask(ctx, txTask.GetID())
+	require.Nil(t, err)
+	require.Equal(t, addrs[0].String(), task1.GetCreator())
+	require.Equal(t, businessHash[:], task1.GetID())
+	_, validTime1 := task1.GetValidTime()
+	require.Equal(t, expiration1, validTime1)
+
+	businessHash2 := sha256.Sum256([]byte("hello world"))
+	expiration2 := time.Now().Add(time.Hour * 2).UTC()
+	txTask2 := types.TxTask{
+		Creator:   addrs[0].String(),
+		TxHash:    businessHash2[:],
+		Bounty:    bounty,
+		ValidTime: expiration2,
+		Status:    types.TaskStatusNil,
+		Score:     100,
+	}
+	require.NoError(t, ok.CreateTask(ctx, addrs[0], &txTask2))
+
+	task2, err := ok.GetTask(ctx, txTask2.GetID())
+	require.Nil(t, err)
+	require.Equal(t, addrs[0].String(), task2.GetCreator())
+	require.Equal(t, businessHash2[:], task2.GetID())
+
+	txTask3 := types.TxTask{
+		Creator:   addrs[0].String(),
+		TxHash:    businessHash2[:],
+		Bounty:    bounty,
+		ValidTime: expiration2,
+		Status:    types.TaskStatusPending,
+		Score:     1,
+	}
+	require.NoError(t, ok.CreateTask(ctx, addrs[0], &txTask3))
+
+	_ = ok.DeleteTask(ctx, task2)
+	_, err = ok.GetTask(ctx, task2.GetID())
+	require.Error(t, err)
 }
